+++ apps/admin-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fluid-ds/admin-react
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @fluid-ds/components@0.1.6
+
 ## 0.0.6
 
 ### Patch Changes